+++ api/logic/load_balancer_logic_test.go
@@ -127,11 +127,11 @@ func TestCreateLoadBalancer(t *testing.T) {
 		IsPublic:       true,
 		Ports:          []models.Port{},
 		HealthCheck:    healthCheck,
-		IdleTimeout:    int64(60),
+		IdleTimeout:    60,
 	}
 
 	testLogic.Backend.EXPECT().
-		CreateLoadBalancer("name", "e1", true, []models.Port{}, healthCheck, int64(60)).
+		CreateLoadBalancer("name", "e1", true, []models.Port{}, healthCheck, 60).
 		Return(retLoadBalancer, nil)
 
 	request := models.CreateLoadBalancerRequest{
@@ -140,7 +140,7 @@ func TestCreateLoadBalancer(t *testing.T) {
 		IsPublic:         true,
 		Ports:            []models.Port{},
 		HealthCheck:      healthCheck,
-		IdleTimeout:      int64(60),
+		IdleTimeout:      60,
 	}
 
 	loadBalancerLogic := NewL0LoadBalancerLogic(testLogic.Logic())
@@ -156,7 +156,7 @@ func TestCreateLoadBalancer(t *testing.T) {
 		IsPublic:         true,
 		Ports:            []models.Port{},
 		HealthCheck:      healthCheck,
-		IdleTimeout:      int64(60),
+		IdleTimeout:      60,
 	}
 
 	testutils.AssertEqual(t, received, expected)
